+++ ariadne/logger.py
@@ -1,16 +1,14 @@
 import logging
 
-from graphql import GraphQLError
-
 from .utils import unwrap_graphql_error
 
 
 logger = logging.getLogger("ariadne")
 
 
-def log_graphql_error(error: GraphQLError, logger):
+def log_error(error: Exception, logger):
     original_error = unwrap_graphql_error(error)
-    if original_error:
-        logger.exception(original_error)
-    else:
-        logger.error(error)
+    if original_error and original_error is not error:
+        error.__suppress_context__ = True
+        error.__cause__ = original_error
+    logger.error(error, exc_info=error)